backend/requirements.txt
# FastAPI & server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Validation & config
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# Auth
python-jose[cryptography]>=3.3.0
bcrypt>=4.1.0
python-multipart>=0.0.6

# DB
sqlalchemy>=2.0.0

# ML
torch>=2.0.0
numpy>=1.24.0
